requirements.txt
bleach==1.5.0
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
decorator==4.1.2
Django==2.0
entrypoints==0.2.3
enum34==1.1.6
gunicorn==19.7.1
html5lib==0.9999999
idna==2.6
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
Markdown==2.6.10
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
numpy==1.13.3
olefile==0.44
pandocfilters==1.4.2
parso==0.1.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
protobuf==3.5.0.post1
psycopg2==2.7.3.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.3
pyzmq==16.0.3
qtconsole==4.3.1
request==0.0.22
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc3
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.13
whitenoise==3.3.1
widgetsnbextension==3.0.8
